i9 refactoring to fix side vs. control issue
New rule is, all ArtificialObjects need to have an order every tick, or else
the simulation won't run; in this way stations can be handled. AI needs to pick
the order, or else the player does.  Facilitate this with more clear &
organized Side enum & Order enum, and a way of validating orders.